Move service status warning indicator to separate component
